+++ sql/catalyst/src/test/scala/org/apache/spark/sql/catalyst/optimizer/ConstantPropagationSuite.scala
@@ -17,7 +17,6 @@
 
 package org.apache.spark.sql.catalyst.optimizer
 
-import org.apache.spark.sql.Row
 import org.apache.spark.sql.catalyst.analysis.EliminateSubqueryAliases
 import org.apache.spark.sql.catalyst.dsl.expressions._
 import org.apache.spark.sql.catalyst.dsl.plans._
@@ -36,28 +35,12 @@ class ConstantPropagationSuite extends PlanTest {
       Batch("AnalysisNodes", Once,
         EliminateSubqueryAliases) ::
         Batch("ConstantPropagation", FixedPoint(10),
-          ConstraintPropagation,
+          ConstantPropagation,
           ConstantFolding,
-          BooleanSimplification,
-          SimplifyBinaryComparison,
-          PruneFilters) :: Nil
+          BooleanSimplification) :: Nil
   }
 
-  val testRelation = LocalRelation('a.int, 'b.int, 'c.int, 'x.boolean, 'y.boolean)
-
-  val data = {
-    val intElements = Seq(null, 1, 2, 3)
-    val booleanElements = Seq(null, true, false)
-    for {
-      a <- intElements
-      b <- intElements
-      c <- intElements
-      x <- booleanElements
-      y <- booleanElements
-    } yield (a, b, c, x, y)
-  }
-
-  val testRelationWithData = LocalRelation.fromExternalRows(testRelation.output, data.map(Row(_)))
+  val testRelation = LocalRelation('a.int, 'b.int, 'c.int)
 
   private val columnA = 'a
   private val columnB = 'b
@@ -171,267 +154,8 @@ class ConstantPropagationSuite extends PlanTest {
       .where(
         columnA === Literal(1) && columnA === Literal(2) && columnB === Add(columnA, Literal(3)))
 
-    val correctAnswer = testRelation.analyze
+    val correctAnswer = testRelation.where(Literal.FalseLiteral).analyze
 
     comparePlans(Optimize.execute(query.analyze), correctAnswer)
-  }
-
-  private def testPropagation(
-      input: Expression,
-      expectEmptyRelation: Boolean,
-      expectedConstraints: Seq[Expression] = Seq.empty) = {
-    val originalQuery = testRelationWithData.where(input).analyze
-    val optimized = Optimize.execute(originalQuery)
-    val correctAnswer = if (expectEmptyRelation) {
-      testRelation
-    } else {
-      testRelationWithData.where(expectedConstraints.reduce(And)).analyze
-    }
-    comparePlans(optimized, correctAnswer)
-  }
-
-  test("Enhanced constraint propagation") {
-    testPropagation('a < 2 && Literal(2) < 'a, true)
-    testPropagation('a < 2 && Literal(2) <= 'a, true)
-    testPropagation('a < 2 && Literal(2) === 'a, true)
-    testPropagation('a < 2 && Literal(2) <=> 'a, true)
-    testPropagation('a < 2 && Literal(2) >= 'a, false, Seq('a < 2))
-    testPropagation('a < 2 && Literal(2) > 'a, false, Seq('a < 2))
-    testPropagation('a <= 2 && Literal(2) < 'a, true)
-    testPropagation('a <= 2 && Literal(2) <= 'a, false, Seq('a === 2))
-    testPropagation('a <= 2 && Literal(2) === 'a, false, Seq('a === 2))
-    testPropagation('a <= 2 && Literal(2) <=> 'a, false, Seq('a === 2))
-    testPropagation('a <= 2 && Literal(2) >= 'a, false, Seq('a <= 2))
-    testPropagation('a <= 2 && Literal(2) > 'a, false, Seq('a < 2))
-    testPropagation('a === 2 && Literal(2) < 'a, true)
-    testPropagation('a === 2 && Literal(2) <= 'a, false, Seq('a === 2))
-    testPropagation('a === 2 && Literal(2) === 'a, false, Seq('a === 2))
-    testPropagation('a === 2 && Literal(2) <=> 'a, false, Seq('a === 2))
-    testPropagation('a === 2 && Literal(2) >= 'a, false, Seq('a === 2))
-    testPropagation('a === 2 && Literal(2) > 'a, true)
-    testPropagation('a <=> 2 && Literal(2) < 'a, true)
-    testPropagation('a <=> 2 && Literal(2) <= 'a, false, Seq('a === 2))
-    testPropagation('a <=> 2 && Literal(2) === 'a, false, Seq('a === 2))
-    testPropagation('a <=> 2 && Literal(2) <=> 'a, false, Seq('a === 2))
-    testPropagation('a <=> 2 && Literal(2) >= 'a, false, Seq('a === 2))
-    testPropagation('a <=> 2 && Literal(2) > 'a, true)
-    testPropagation('a >= 2 && Literal(2) < 'a, false, Seq(Literal(2) < 'a))
-    testPropagation('a >= 2 && Literal(2) <= 'a, false, Seq(Literal(2) <= 'a))
-    testPropagation('a >= 2 && Literal(2) === 'a, false, Seq('a === 2))
-    testPropagation('a >= 2 && Literal(2) <=> 'a, false, Seq('a === 2))
-    testPropagation('a >= 2 && Literal(2) >= 'a, false, Seq('a === 2))
-    testPropagation('a >= 2 && Literal(2) > 'a, true)
-    testPropagation('a > 2 && Literal(2) < 'a, false, Seq(Literal(2) < 'a))
-    testPropagation('a > 2 && Literal(2) <= 'a, false, Seq(Literal(2) < 'a))
-    testPropagation('a > 2 && Literal(2) === 'a, true)
-    testPropagation('a > 2 && Literal(2) <=> 'a, true)
-    testPropagation('a > 2 && Literal(2) >= 'a, true)
-    testPropagation('a > 2 && Literal(2) > 'a, true)
-
-    testPropagation('a < 2 && Literal(3) < 'a, true)
-    testPropagation('a < 2 && Literal(3) <= 'a, true)
-    testPropagation('a < 2 && Literal(3) === 'a, true)
-    testPropagation('a < 2 && Literal(3) <=> 'a, true)
-    testPropagation('a < 2 && Literal(3) >= 'a, false, Seq('a < 2))
-    testPropagation('a < 2 && Literal(3) > 'a, false, Seq('a < 2))
-    testPropagation('a <= 2 && Literal(3) < 'a, true)
-    testPropagation('a <= 2 && Literal(3) <= 'a, true)
-    testPropagation('a <= 2 && Literal(3) === 'a, true)
-    testPropagation('a <= 2 && Literal(3) <=> 'a, true)
-    testPropagation('a <= 2 && Literal(3) >= 'a, false, Seq('a <= 2))
-    testPropagation('a <= 2 && Literal(3) > 'a, false, Seq('a <= 2))
-    testPropagation('a === 2 && Literal(3) < 'a, true)
-    testPropagation('a === 2 && Literal(3) <= 'a, true)
-    testPropagation('a === 2 && Literal(3) === 'a, true)
-    testPropagation('a === 2 && Literal(3) <=> 'a, true)
-    testPropagation('a === 2 && Literal(3) >= 'a, false, Seq('a === 2))
-    testPropagation('a === 2 && Literal(3) > 'a, false, Seq('a === 2))
-    testPropagation('a <=> 2 && Literal(3) < 'a, true)
-    testPropagation('a <=> 2 && Literal(3) <= 'a, true)
-    testPropagation('a <=> 2 && Literal(3) === 'a, true)
-    testPropagation('a <=> 2 && Literal(3) <=> 'a, true)
-    testPropagation('a <=> 2 && Literal(3) >= 'a, false, Seq('a === 2))
-    testPropagation('a <=> 2 && Literal(3) > 'a, false, Seq('a === 2))
-    testPropagation('a >= 2 && Literal(3) < 'a, false, Seq(Literal(3) < 'a))
-    testPropagation('a >= 2 && Literal(3) <= 'a, false, Seq(Literal(3) <= 'a))
-    testPropagation('a >= 2 && Literal(3) === 'a, false, Seq(Literal(3) === 'a))
-    testPropagation('a >= 2 && Literal(3) <=> 'a, false, Seq(Literal(3) === 'a))
-    testPropagation('a >= 2 && Literal(3) >= 'a, false, Seq(Literal(2) <= 'a, 'a <= Literal(3)))
-    testPropagation('a >= 2 && Literal(3) > 'a, false, Seq(Literal(2) <= 'a, 'a < Literal(3)))
-    testPropagation('a > 2 && Literal(3) < 'a, false, Seq(Literal(3) < 'a))
-    testPropagation('a > 2 && Literal(3) <= 'a, false, Seq(Literal(3) <= 'a))
-    testPropagation('a > 2 && Literal(3) === 'a, false, Seq(Literal(3) === 'a))
-    testPropagation('a > 2 && Literal(3) <=> 'a, false, Seq(Literal(3) === 'a))
-    testPropagation('a > 2 && Literal(3) >= 'a, false, Seq(Literal(2) < 'a, 'a <= Literal(3)))
-    testPropagation('a > 2 && Literal(3) > 'a, false, Seq(Literal(2) < 'a, 'a < Literal(3)))
-
-    testPropagation('a < 3 && Literal(2) < 'a, false, Seq('a < 3 && Literal(2) < 'a))
-    testPropagation('a < 3 && Literal(2) <= 'a, false, Seq('a < 3 && Literal(2) <= 'a))
-    testPropagation('a < 3 && Literal(2) === 'a, false, Seq('a === 2))
-    testPropagation('a < 3 && Literal(2) <=> 'a, false, Seq('a === 2))
-    testPropagation('a < 3 && Literal(2) >= 'a, false, Seq('a <= 2))
-    testPropagation('a < 3 && Literal(2) > 'a, false, Seq('a < 2))
-    testPropagation('a <= 3 && Literal(2) < 'a, false, Seq('a <= 3 && Literal(2) < 'a))
-    testPropagation('a <= 3 && Literal(2) <= 'a, false, Seq('a <= 3 && Literal(2) <= 'a))
-    testPropagation('a <= 3 && Literal(2) === 'a, false, Seq('a === 2))
-    testPropagation('a <= 3 && Literal(2) <=> 'a, false, Seq('a === 2))
-    testPropagation('a <= 3 && Literal(2) >= 'a, false, Seq('a <= 2))
-    testPropagation('a <= 3 && Literal(2) > 'a, false, Seq('a < 2))
-    testPropagation('a === 3 && Literal(2) < 'a, false, Seq('a === 3))
-    testPropagation('a === 3 && Literal(2) <= 'a, false, Seq('a === 3))
-    testPropagation('a === 3 && Literal(2) === 'a, true)
-    testPropagation('a === 3 && Literal(2) <=> 'a, true)
-    testPropagation('a === 3 && Literal(2) >= 'a, true)
-    testPropagation('a === 3 && Literal(2) > 'a, true)
-    testPropagation('a <=> 3 && Literal(2) < 'a, false, Seq('a === 3))
-    testPropagation('a <=> 3 && Literal(2) <= 'a, false, Seq('a === 3))
-    testPropagation('a <=> 3 && Literal(2) === 'a, true)
-    testPropagation('a <=> 3 && Literal(2) <=> 'a, true)
-    testPropagation('a <=> 3 && Literal(2) >= 'a, true)
-    testPropagation('a <=> 3 && Literal(2) > 'a, true)
-    testPropagation('a >= 3 && Literal(2) < 'a, false, Seq(Literal(3) <= 'a))
-    testPropagation('a >= 3 && Literal(2) <= 'a, false, Seq(Literal(3) <= 'a))
-    testPropagation('a >= 3 && Literal(2) === 'a, true)
-    testPropagation('a >= 3 && Literal(2) <=> 'a, true)
-    testPropagation('a >= 3 && Literal(2) >= 'a, true)
-    testPropagation('a >= 3 && Literal(2) > 'a, true)
-    testPropagation('a > 3 && Literal(2) < 'a, false, Seq(Literal(3) < 'a))
-    testPropagation('a > 3 && Literal(2) <= 'a, false, Seq(Literal(3) < 'a))
-    testPropagation('a > 3 && Literal(2) === 'a, true)
-    testPropagation('a > 3 && Literal(2) <=> 'a, true)
-    testPropagation('a > 3 && Literal(2) >= 'a, true)
-    testPropagation('a > 3 && Literal(2) > 'a, true)
-
-    testPropagation('a < 'b && 'b < 'a, true)
-    testPropagation('a < 'b && 'b <= 'a, true)
-    testPropagation('a < 'b && 'b === 'a, true)
-    testPropagation('a < 'b && 'b <=> 'a, true)
-    testPropagation('a < 'b && 'b >= 'a, false, Seq('a < 'b))
-    testPropagation('a < 'b && 'b > 'a, false, Seq('a < 'b))
-    testPropagation('a <= 'b && 'b < 'a, true)
-    testPropagation('a <= 'b && 'b <= 'a, false, Seq('a === 'b))
-    testPropagation('a <= 'b && 'b === 'a, false, Seq('a === 'b))
-    testPropagation('a <= 'b && 'b <=> 'a, false, Seq('a === 'b))
-    testPropagation('a <= 'b && 'b >= 'a, false, Seq('a <= 'b))
-    testPropagation('a <= 'b && 'b > 'a, false, Seq('a < 'b))
-    testPropagation('a === 'b && 'b < 'a, true)
-    testPropagation('a === 'b && 'b <= 'a, false, Seq('a === 'b))
-    testPropagation('a === 'b && 'b === 'a, false, Seq('a === 'b))
-    testPropagation('a === 'b && 'b <=> 'a, false, Seq('a === 'b))
-    testPropagation('a === 'b && 'b >= 'a, false, Seq('a === 'b))
-    testPropagation('a === 'b && 'b > 'a, true)
-    testPropagation('a <=> 'b && 'b < 'a, true)
-    testPropagation('a <=> 'b && 'b <= 'a, false, Seq('a === 'b))
-    testPropagation('a <=> 'b && 'b === 'a, false, Seq('a === 'b))
-    testPropagation('a <=> 'b && 'b <=> 'a, false, Seq('a <=> 'b))
-    testPropagation('a <=> 'b && 'b >= 'a, false, Seq('a === 'b))
-    testPropagation('a <=> 'b && 'b > 'a, true)
-    testPropagation('a >= 'b && 'b < 'a, false, Seq('b < 'a))
-    testPropagation('a >= 'b && 'b <= 'a, false, Seq('b <= 'a))
-    testPropagation('a >= 'b && 'b === 'a, false, Seq('a === 'b))
-    testPropagation('a >= 'b && 'b <=> 'a, false, Seq('a === 'b))
-    testPropagation('a >= 'b && 'b >= 'a, false, Seq('a === 'b))
-    testPropagation('a >= 'b && 'b > 'a, true)
-    testPropagation('a > 'b && 'b < 'a, false, Seq('b < 'a))
-    testPropagation('a > 'b && 'b <= 'a, false, Seq('b < 'a))
-    testPropagation('a > 'b && 'b === 'a, true)
-    testPropagation('a > 'b && 'b <=> 'a, true)
-    testPropagation('a > 'b && 'b >= 'a, true)
-    testPropagation('a > 'b && 'b > 'a, true)
-
-    testPropagation('a < abs('b) && abs('b) < 'a, true)
-    testPropagation('a < abs('b) && abs('b) <= 'a, true)
-    testPropagation('a < abs('b) && abs('b) === 'a, true)
-    testPropagation('a < abs('b) && abs('b) <=> 'a, true)
-    testPropagation('a < abs('b) && abs('b) >= 'a, false, Seq('a < abs('b)))
-    testPropagation('a < abs('b) && abs('b) > 'a, false, Seq('a < abs('b)))
-    testPropagation('a <= abs('b) && abs('b) < 'a, true)
-    testPropagation('a <= abs('b) && abs('b) <= 'a, false, Seq('a === abs('b)))
-    testPropagation('a <= abs('b) && abs('b) === 'a, false, Seq('a === abs('b)))
-    testPropagation('a <= abs('b) && abs('b) <=> 'a, false, Seq('a === abs('b)))
-    testPropagation('a <= abs('b) && abs('b) >= 'a, false, Seq('a <= abs('b)))
-    testPropagation('a <= abs('b) && abs('b) > 'a, false, Seq('a < abs('b)))
-    testPropagation('a === abs('b) && abs('b) < 'a, true)
-    testPropagation('a === abs('b) && abs('b) <= 'a, false, Seq('a === abs('b)))
-    testPropagation('a === abs('b) && abs('b) === 'a, false, Seq('a === abs('b)))
-    testPropagation('a === abs('b) && abs('b) <=> 'a, false, Seq('a === abs('b)))
-    testPropagation('a === abs('b) && abs('b) >= 'a, false, Seq('a === abs('b)))
-    testPropagation('a === abs('b) && abs('b) > 'a, true)
-    testPropagation('a <=> abs('b) && abs('b) < 'a, true)
-    testPropagation('a <=> abs('b) && abs('b) <= 'a, false, Seq('a === abs('b)))
-    testPropagation('a <=> abs('b) && abs('b) === 'a, false, Seq('a === abs('b)))
-    testPropagation('a <=> abs('b) && abs('b) <=> 'a, false, Seq('a <=> abs('b)))
-    testPropagation('a <=> abs('b) && abs('b) >= 'a, false, Seq('a === abs('b)))
-    testPropagation('a <=> abs('b) && abs('b) > 'a, true)
-    testPropagation('a >= abs('b) && abs('b) < 'a, false, Seq(abs('b) < 'a))
-    testPropagation('a >= abs('b) && abs('b) <= 'a, false, Seq(abs('b) <= 'a))
-    testPropagation('a >= abs('b) && abs('b) === 'a, false, Seq('a === abs('b)))
-    testPropagation('a >= abs('b) && abs('b) <=> 'a, false, Seq('a === abs('b)))
-    testPropagation('a >= abs('b) && abs('b) >= 'a, false, Seq('a === abs('b)))
-    testPropagation('a >= abs('b) && abs('b) > 'a, true)
-    testPropagation('a > abs('b) && abs('b) < 'a, false, Seq(abs('b) < 'a))
-    testPropagation('a > abs('b) && abs('b) <= 'a, false, Seq(abs('b) < 'a))
-    testPropagation('a > abs('b) && abs('b) === 'a, true)
-    testPropagation('a > abs('b) && abs('b) <=> 'a, true)
-    testPropagation('a > abs('b) && abs('b) >= 'a, true)
-    testPropagation('a > abs('b) && abs('b) > 'a, true)
-
-    testPropagation(('x || 'a < abs('b) || 'y) && abs('b) < 'a, false, Seq('x || 'y, abs('b) < 'a))
-    testPropagation(('x || 'a < abs('b) || 'y) && abs('b) <= 'a, false,
-      Seq('x || 'y, abs('b) <= 'a))
-    testPropagation(('x || 'a < abs('b) || 'y) && abs('b) === 'a, false,
-      Seq('x || 'y, abs('b) === 'a))
-    testPropagation(('x || 'a < abs('b) || 'y) && abs('b) <=> 'a, false,
-      Seq('x || 'y, abs('b) <=> 'a))
-    testPropagation(('x || 'a < abs('b) || 'y) && abs('b) >= 'a, false,
-      Seq('x || 'a < abs('b) || 'y, 'a <= abs('b)))
-    testPropagation(('x || 'a < abs('b) || 'y) && abs('b) > 'a, false, Seq('a < abs('b)))
-    testPropagation(('x || 'a <= abs('b) || 'y) && abs('b) < 'a, false, Seq('x || 'y, abs('b) < 'a))
-    testPropagation(('x || 'a <= abs('b) || 'y) && abs('b) <= 'a, false,
-      Seq('x || 'a === abs('b) || 'y, abs('b) <= 'a))
-    testPropagation(('x || 'a <= abs('b) || 'y) && abs('b) === 'a, false, Seq(abs('b) === 'a))
-    testPropagation(('x || 'a <= abs('b) || 'y) && abs('b) <=> 'a, false,
-      Seq('x || 'a === abs('b) || 'y, abs('b) <=> 'a))
-    testPropagation(('x || 'a <= abs('b) || 'y) && abs('b) >= 'a, false, Seq('a <= abs('b)))
-    testPropagation(('x || 'a <= abs('b) || 'y) && abs('b) > 'a, false, Seq('a < abs('b)))
-    testPropagation(('x || 'a === abs('b) || 'y) && abs('b) < 'a, false,
-      Seq('x || 'y, abs('b) < 'a))
-    testPropagation(('x || 'a === abs('b) || 'y) && abs('b) <= 'a, false,
-      Seq('x || 'a === abs('b) || 'y, abs('b) <= 'a))
-    testPropagation(('x || 'a === abs('b) || 'y) && abs('b) === 'a, false, Seq(abs('b) === 'a))
-    testPropagation(('x || 'a === abs('b) || 'y) && abs('b) <=> 'a, false,
-      Seq('x || 'a === abs('b) || 'y, abs('b) <=> 'a))
-    testPropagation(('x || 'a === abs('b) || 'y) && abs('b) >= 'a, false,
-      Seq('x || 'a === abs('b) || 'y, 'a <= abs('b)))
-    testPropagation(('x || 'a === abs('b) || 'y) && abs('b) > 'a, false,
-      Seq('x || 'y, 'a < abs('b)))
-    testPropagation(('x || 'a <=> abs('b) || 'y) && abs('b) < 'a, false,
-      Seq('x || 'y, abs('b) < 'a))
-    testPropagation(('x || 'a <=> abs('b) || 'y) && abs('b) <= 'a, false,
-      Seq('x || 'a === abs('b) || 'y, abs('b) <= 'a))
-    testPropagation(('x || 'a <=> abs('b) || 'y) && abs('b) === 'a, false, Seq(abs('b) === 'a))
-    testPropagation(('x || 'a <=> abs('b) || 'y) && abs('b) <=> 'a, false, Seq(abs('b) <=> 'a))
-    testPropagation(('x || 'a <=> abs('b) || 'y) && abs('b) >= 'a, false,
-      Seq('x || 'a === abs('b) || 'y, 'a <= abs('b)))
-    testPropagation(('x || 'a <=> abs('b) || 'y) && abs('b) > 'a, false,
-      Seq('x || 'y, 'a < abs('b)))
-    testPropagation(('x || 'a >= abs('b) || 'y) && abs('b) < 'a, false, Seq(abs('b) < 'a))
-    testPropagation(('x || 'a >= abs('b) || 'y) && abs('b) <= 'a, false, Seq(abs('b) <= 'a))
-    testPropagation(('x || 'a >= abs('b) || 'y) && abs('b) === 'a, false, Seq(abs('b) === 'a))
-    testPropagation(('x || 'a >= abs('b) || 'y) && abs('b) <=> 'a, false,
-      Seq('x || 'a === abs('b) || 'y, abs('b) <=> 'a))
-    testPropagation(('x || 'a >= abs('b) || 'y) && abs('b) >= 'a, false,
-      Seq('x || 'a === abs('b) || 'y, 'a <= abs('b)))
-    testPropagation(('x || 'a >= abs('b) || 'y) && abs('b) > 'a, false, Seq('x || 'y, 'a < abs('b)))
-    testPropagation(('x || 'a > abs('b) || 'y) && abs('b) < 'a, false, Seq(abs('b) < 'a))
-    testPropagation(('x || 'a > abs('b) || 'y) && abs('b) <= 'a, false,
-      Seq('x || abs('b) < 'a || 'y, abs('b) <= 'a))
-    testPropagation(('x || 'a > abs('b) || 'y) && abs('b) === 'a, false,
-      Seq('x || 'y, abs('b) === 'a))
-    testPropagation(('x || 'a > abs('b) || 'y) && abs('b) <=> 'a, false,
-      Seq('x || 'y, abs('b) <=> 'a))
-    testPropagation(('x || 'a > abs('b) || 'y) && abs('b) >= 'a, false,
-      Seq('x || 'y, 'a <= abs('b)))
-    testPropagation(('x || 'a > abs('b) || 'y) && abs('b) > 'a, false, Seq('x || 'y, 'a < abs('b)))
   }
 }
